fix: inline verified migration reader binding
